Change to use hawx/mux and hawx/route
This meant changing a few routes, and a bit of hackiness. Thing is
hawx/route, like the httprouter it is based on, does not allow

  /something
  /:wildcard

route registrations, once a wildcard is in a position that is all there
can be. So I've added a couple of subrouters for the "-" and "assets"
""namespaces"". Seems ok. Works at least.

Also rearranged /edit/whatevs to /whatevs/edit, sim. for delete.